backend/requirements.txt
pandas
tqdm
fastapi
uvicorn[standard]
python-dotenv
pinecone
google-genai
requests
beautifulsoup4
ratelimit
backoff
neo4j
python-multipart
pyjwt
motor
pymongo
authlib
httpx
cryptography
